+++ blacksmith/experiments/torch/albert/test_albert_finetuning.py
@@ -18,39 +18,7 @@ from blacksmith.tools.cli import generate_config, parse_cli_options
 from blacksmith.tools.device_manager import DeviceManager
 from blacksmith.tools.logging_manager import TrainingLogger
 from blacksmith.tools.reproducibility_manager import ReproducibilityManager
-from blacksmith.tools.workaround_utils import TTLayerNorm
-
-
-def replace_layernorm(module):
-    """
-    Recursively replaces all nn.LayerNorm modules in a PyTorch model
-    with CustomLayerNorm, preserving their weights and biases.
-    """
-    for name, child in module.named_children():
-        if isinstance(child, torch.nn.LayerNorm):
-            # 1. Initialize the custom layer with the same configuration
-            custom_ln = TTLayerNorm(
-                normalized_shape=child.normalized_shape,
-                eps=child.eps,
-                elementwise_affine=child.elementwise_affine,
-                device=child.weight.device if child.weight is not None else None,
-                dtype=child.weight.dtype if child.weight is not None else None,
-            )
-
-            # 2. Copy the learned parameters if elementwise_affine is True
-            if child.elementwise_affine:
-                with torch.no_grad():
-                    custom_ln.weight.copy_(child.weight)
-                    custom_ln.bias.copy_(child.bias)
-
-            # 3. Replace the original layer with the custom one
-            setattr(module, name, custom_ln)
-        else:
-            # Recursively apply to child modules (e.g., inside nn.Sequential)
-            replace_layernorm(child)
-
-    return module
-
+from blacksmith.tools.workaround_utils import replace_layernorm
 
 def validate(
     model: torch.nn.Module,
