+++ blacksmith/tools/workaround_utils.py
@@ -40,11 +40,10 @@ def transform_labels(labels, ignored_index, vocab_size):
     return expected_output, labels_mask
 
 
-# Custom LayerNorm for TT-Forge. Necessary because of Albert experiment getting inf loss.
 class TTLayerNorm(nn.Module):
     def __init__(self, normalized_shape, eps=1e-5, elementwise_affine=True, device=None, dtype=None):
         super().__init__()
-        # Handle cases where normalized_shape is an int
+        # Handle cases where `normalized_shape` is an int.
         if isinstance(normalized_shape, int):
             normalized_shape = (normalized_shape,)
         self.normalized_shape = tuple(normalized_shape)
@@ -59,19 +58,52 @@ class TTLayerNorm(nn.Module):
             self.register_parameter("bias", None)
 
     def forward(self, x):
-        # Determine which dimensions to compute the mean and variance over
+        # Determine which dimensions to compute the mean and variance over.
         dims = tuple(range(-len(self.normalized_shape), 0))
 
-        # Compute mean and variance
+        # Compute mean and variance.
         mean = x.mean(dim=dims, keepdim=True)
-        # We use unbiased=False to match PyTorch's native nn.LayerNorm implementation
+        # We use `unbiased=False` to match PyTorch's native `nn.LayerNorm` implementation.
         var = x.var(dim=dims, unbiased=False, keepdim=True)
 
-        # Normalize
+        # Normalize.
         x_norm = (x - mean) / torch.sqrt(var + self.eps)
 
-        # Apply learnable affine parameters if specified
+        # Apply learnable affine parameters if specified.
         if self.elementwise_affine:
             x_norm = x_norm * self.weight + self.bias
 
         return x_norm
+
+
+def replace_layernorm(module):
+    """
+    Recursively replaces all nn.LayerNorm modules in a PyTorch model
+    with CustomLayerNorm, preserving their weights and biases.
+    """
+    for name, child in module.named_children():
+        if isinstance(child, torch.nn.LayerNorm):
+            # 1. Initialize the custom layer with the same configuration.
+            custom_ln = TTLayerNorm(
+                normalized_shape=child.normalized_shape,
+                eps=child.eps,
+                elementwise_affine=child.elementwise_affine,
+                device=child.weight.device if child.weight is not None else None,
+                dtype=child.weight.dtype if child.weight is not None else None,
+            )
+
+            # 2. Copy the learned parameters if `elementwise_affine` is True.
+            if child.elementwise_affine:
+                with torch.no_grad():
+                    custom_ln.weight.copy_(child.weight)
+                    custom_ln.bias.copy_(child.bias)
+                    custom_ln.weight.requires_grad = False
+                    custom_ln.bias.requires_grad = False
+
+            # 3. Replace the original layer with the custom one.
+            setattr(module, name, custom_ln)
+        else:
+            # Recursively apply to child modules (e.g., inside `nn.Sequential`).
+            replace_layernorm(child)
+
+    return module
